==> 1. data/2. non_fruitfly_data/1. swc_to_points_segments.ipynb <==
import pandas as pd

# Define the name for the output files and specify the source and destination paths
name = 'swc_skeleton'
path_source = '1. raw_data/'
path_save = '2. points_segments/'

# Read the SWC file into a DataFrame. SWC files are space-delimited.
df = pd.read_csv(path_source + '04546925828.swc', delimiter=' ')

# Extract segment information from the DataFrame
# The 'Parent' column specifies the ID of the parent point for a segment, and 'PointNo' is the ID of the current point.
segments = df[['Parent', 'PointNo']]
# Rename the columns for clarity
segments = segments.rename(columns={'Parent': 'pt_id1', 'PointNo': 'pt_id2'})
# Remove rows where either 'pt_id1' or 'pt_id2' is -1, indicating no connection or an invalid point
segments = segments[(segments['pt_id1'] != -1) & (segments['pt_id2'] != -1)]
# Reset the index of the DataFrame, dropping the old index
segments = segments.reset_index(drop=True)
# Rename the index column as 'seg_id' to uniquely identify each segment
segments = segments.rename(columns={'level_0': 'seg_id'})
# Save the segments information to a CSV file
segments.to_csv(path_save + name + '.segments.csv')

# Extract point information from the DataFrame
# Includes point ID, coordinates (X, Y, Z), and Radius
points = df[['PointNo', 'X', 'Y', 'Z', 'Radius']].copy()
# Rename columns for clarity and consistency
points = points.rename(columns={'PointNo': 'pt_id', 'X': 'x', 'Y': 'y', 'Z': 'z', 'Radius': 'radius'})
# Calculate the diameter of each point (neuron part) as twice the radius
points['diameter'] = points['radius'] * 2
# Reset the index of the DataFrame, ensuring a clean, sequential index
points = points.reset_index(drop=True)
# Save the points information to a CSV file
points.to_csv(path_save + name + '.points.csv')

points.head()

segments.head()
